subscription/requirements.txt
Flask==2.2.2
stripe==12.0.0
scipy==1.12.0
numpy==1.23.5
mysql-connector-python==9.2.0